Completed p1 and refactored p2,p3. have not added stuff for p2,p3
Put partials in folders

On principle3, I encountered an issue involving the explanation. Since I saw
that it was the only explanation with html in it (the sup tag), I decided to
put an if statement in the default.html so that we can keep using the
jumbotron in the layout and we can deal with the special case. That does mean
that principle 3 is the only one without an explanation in its front matter.