Added All the Pages
Created all of the principle pages and all of them follow the structure of the
first principle page. Created all of the partial pages as well.

I made some small edits to the img names and editted some more of p1.

I have not added anything new in terms of principle partial content.